draw the rest of the owl
